+++ MatrixKit/Controllers/MXKAuthenticationViewController.h
@@ -80,6 +80,7 @@
 @property (weak, nonatomic) IBOutlet UILabel *homeServerLabel;
 @property (weak, nonatomic) IBOutlet UITextField *homeServerTextField;
 @property (weak, nonatomic) IBOutlet UILabel *homeServerInfoLabel;
+@property (weak, nonatomic) IBOutlet UIView *identityServerContainer;
 @property (weak, nonatomic) IBOutlet UILabel *identityServerLabel;
 @property (weak, nonatomic) IBOutlet UITextField *identityServerTextField;
 @property (weak, nonatomic) IBOutlet UILabel *identityServerInfoLabel;
@@ -231,6 +232,12 @@
  @param identityServerUrl the identity server url to use
  */
 - (void)setIdentityServerTextFieldText:(NSString *)identityServerUrl;
+
+/**
+ Fetch the identity server from the wellknown API of the selected homeserver.
+ and check if the HS requires an identity server.
+ */
+- (void)checkIdentityServer;
 
 /**
  Force dismiss keyboard