+++ CHANGES.rst
@@ -1,8 +1,13 @@
-Changes in MatrixKit in 0.10.3 (2019-xx-xx)
+Changes in MatrixKit in 0.11.0 (2019-10-)
 ==========================================
 
 Improvements:
- *
+ * Upgrade MatrixSDK version ([v0.14.0](https://github.com/matrix-org/matrix-ios-sdk/releases/tag/v0.14.0)).
+ * MXKDeviceView: Make clear that device names are publicly readable (vector-im/riot-ios/issues/2662).
+ * Privacy: Remove the bind true flag from 3PID adds in settings (vector-im/riot-ios/issues/2650).
+ * Privacy: Remove the ability to set an IS at login/registration (vector-im/riot-ios/issues/2661).
+ * Privacy: Use wellknown to discover the IS of a custom HS (vector-im/riot-ios/issues/2686).
+ * Tools: Add human readable MSISDN formatting method.
 
 Bug fix:
  * Display correctly the revoked third-party invite.
@@ -15,6 +20,8 @@ Improvements:
  * Support soft logout (vector-im/riot-ios/issues/2540).
  * MXKRoomBubbleCellData: Add method to get bubble component index from event id.
  * MXKEmail: force in lowercase the email address.
+ * Use MXIdentityService to perform identity server requests (vector-im/riot-ios#2647).
+ * Support identity server v2 API (vector-im/riot-ios#2603 and /vector-im/riot-ios#2652).
 
  Bug fix:
  * APNS Push: fix logic when enabling APNS push. Avoid calling nil callback method.